+++ SandboxiePlus/SandMan/OnlineUpdater.cpp
@@ -417,7 +417,7 @@ bool COnlineUpdater::AskDownload(const QVariantMap& Data)
 	mb.setCheckBoxVisible(m_CheckMode != eManual);
 
 	if (!UpdateUrl.isEmpty() || !DownloadUrl.isEmpty() || Data.contains("files")) {
-		mb.setStandardButtons(QDialogButtonBox::Yes | QDialogButtonBox::No);
+		mb.setStandardButtons(QDialogButtonBox::Yes | QDialogButtonBox::No | QDialogButtonBox::Cancel);
 		mb.setDefaultButton(QDialogButtonBox::Yes);
 	}
 	else
@@ -434,8 +434,16 @@ bool COnlineUpdater::AskDownload(const QVariantMap& Data)
 		else
 			QDesktopServices::openUrl(UpdateUrl);
 	}
-	else if (mb.isChecked())
-		theConf->SetValue("Options/IgnoredUpdates", m_IgnoredUpdates << VersionStr);
+	else 
+	{
+		if (mb.clickedStandardButton() == QDialogButtonBox::Cancel) {
+			theConf->SetValue("Updater/PendingUpdate", ""); 
+			theGUI->UpdateLabel();
+		}
+
+		if (mb.isChecked())
+			theConf->SetValue("Options/IgnoredUpdates", m_IgnoredUpdates << VersionStr);
+	}
 	return false;
 }