+++ SandboxiePlus/SandMan/SandMan.cpp
@@ -1432,7 +1432,7 @@ void CSandMan::timerEvent(QTimerEvent* pEvent)
 
 		bool bUpdatePending = !theConf->GetString("Updater/PendingUpdate").isEmpty();
 
-		if (m_bIconEmpty != (ActiveProcesses == 0) || m_bIconBusy != bIconBusy || m_iIconDisabled != (bForceProcessDisabled ? 1 : 0) || bUpdatePending)
+		if (m_bIconEmpty != (ActiveProcesses == 0) || m_bIconBusy != bIconBusy || m_iIconDisabled != (bForceProcessDisabled ? 1 : 0) || bUpdatePending || m_bIconSun)
 		{
 			m_bIconEmpty = (ActiveProcesses == 0);
 			m_bIconBusy = bIconBusy;